Set units to blocked status when scaling beyond 1
Instead of throwing an error, block the units when scaled beyond 1.

Throwing an error causes the charm to go into a bad state, where it
can't even be removed without manual intervention. This even causes an
issue with a single unit, as sometimes leader status is removed before
the charm has finished processing all events.